Refactoring
Moving recurrent functions to services in order to avoid duplicate functions doing the same thing.
