+++ src/pages/occurrences/occurrences.ts
@@ -3,6 +3,7 @@ import { IonicPage, NavController, NavParams, ModalController, App, Platform, Vi
 import { SemanticDataService } from '../../app/services/semantic-data/semantic-data.service';
 import { ConfigService } from '@ngx-config/core';
 import { TextService } from '../../app/services/texts/text.service';
+import { TooltipService } from '../../app/services/tooltips/tooltip.service';
 import { OccurrenceService } from '../../app/services/occurrence/occurence.service';
 import { LanguageService } from '../../app/services/languages/language.service';
 import { Occurrence, OccurrenceType, OccurrenceResult } from '../../app/models/occurrence.model';
@@ -44,8 +45,6 @@ export class OccurrencesPage {
   source: string = null;
   description: string = null;
   country: string = null;
-  year_born: string = null;
-  year_deceased: string = null;
   year_born_deceased_string: string = null;
   publisher: string = null;
   published_year: string = null;
@@ -80,6 +79,7 @@ export class OccurrencesPage {
               private app: App,
               private platform: Platform,
               protected textService: TextService,
+              private tooltipService: TooltipService,
               public translate: TranslateService,
               public occurrenceService: OccurrenceService,
               public viewCtrl: ViewController,
@@ -130,33 +130,8 @@ export class OccurrencesPage {
       this.authors = [];
     }
 
-    // Get the born and deceased years without leading zeros and possible 'BC' indicators
-    this.year_born = (this.occurrenceResult.date_born !== undefined && this.occurrenceResult.date_born !== null) ?
-                              String(this.occurrenceResult.date_born).split('-')[0].replace(/^0+/, '').split(' ')[0] : null;
-    this.year_deceased = (this.occurrenceResult.date_deceased !== undefined && this.occurrenceResult.date_deceased !== null) ?
-                              String(this.occurrenceResult.date_deceased).split('-')[0].replace(/^0+/, '').split(' ')[0] : null;
-    // Get translation for 'BC'
-    let bcTranslation = 'BC';
-    this.translate.get('BC').subscribe(
-      translation => {
-        bcTranslation = translation;
-      }, error => { }
-    );
-    const bcIndicatorDeceased = (String(this.occurrenceResult.date_deceased).includes('BC')) ? ' ' + bcTranslation : '';
-    let bcIndicatorBorn = (String(this.occurrenceResult.date_born).includes('BC')) ? ' ' + bcTranslation : '';
-    if (String(this.occurrenceResult.date_born).includes('BC') && bcIndicatorDeceased === bcIndicatorBorn) {
-      // Born and deceased are both BC --> don't add indicator to year born
-      bcIndicatorBorn = '';
-    }
-    // Construct string with year born and year deceased for output
-    this.year_born_deceased_string = '';
-    if (this.year_born !== null && this.year_deceased !== null && this.year_born !== 'null' && this.year_deceased !== 'null') {
-      this.year_born_deceased_string += '(' + this.year_born + bcIndicatorBorn + '–' + this.year_deceased + bcIndicatorDeceased + ')';
-    } else if (this.year_born !== null && this.year_born !== 'null') {
-      this.year_born_deceased_string += '(* ' + this.year_born + bcIndicatorBorn + ')';
-    } else if (this.year_deceased !== null && this.year_deceased !== 'null') {
-      this.year_born_deceased_string += '(&#8224; ' + this.year_deceased + bcIndicatorDeceased + ')';
-    }
+    this.year_born_deceased_string = this.tooltipService.constructYearBornDeceasedString(this.occurrenceResult.date_born,
+      this.occurrenceResult.date_deceased);
 
     try {
       this.singleOccurrenceType = this.config.getSettings('SingleOccurrenceType');
